+++ besu/src/test/java/org/hyperledger/besu/cli/options/TransactionPoolOptionsTest.java
@@ -69,6 +69,42 @@ public class TransactionPoolOptionsTest
   }
 
   @Test
+  public void senderLimitedTxPool_derived() {
+    final TestBesuCommand cmd = parseCommand("--tx-pool-limit-by-account-percentage=0.002");
+
+    final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
+    final TransactionPoolConfiguration config = options.toDomainObject().build();
+    assertThat(config.getTxPoolMaxFutureTransactionByAccount()).isEqualTo(9);
+
+    assertThat(commandOutput.toString(UTF_8)).isEmpty();
+    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
+  }
+
+  @Test
+  public void senderLimitedTxPoolFloor_derived() {
+    final TestBesuCommand cmd = parseCommand("--tx-pool-limit-by-account-percentage=0.0001");
+
+    final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
+    final TransactionPoolConfiguration config = options.toDomainObject().build();
+    assertThat(config.getTxPoolMaxFutureTransactionByAccount()).isEqualTo(1);
+
+    assertThat(commandOutput.toString(UTF_8)).isEmpty();
+    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
+  }
+
+  @Test
+  public void senderLimitedTxPoolCeiling_violated() {
+    final TestBesuCommand cmd = parseCommand("--tx-pool-limit-by-account-percentage=1.00000001");
+
+    final TransactionPoolOptions options = getOptionsFromBesuCommand(cmd);
+    final TransactionPoolConfiguration config = options.toDomainObject().build();
+
+    assertThat(commandOutput.toString(UTF_8)).isEmpty();
+    assertThat(commandErrorOutput.toString(UTF_8)).isEmpty();
+    assertThat(config).isNotNull();
+  }
+
+  @Test
   public void strictTxReplayProtection_default() {
     final TestBesuCommand cmd = parseCommand();
 
@@ -120,7 +156,8 @@ public class TransactionPoolOptionsTest
     return ImmutableTransactionPoolConfiguration.builder()
         .strictTransactionReplayProtectionEnabled(false)
         .txMessageKeepAliveSeconds(defaultValue.getTxMessageKeepAliveSeconds())
-        .eth65TrxAnnouncedBufferingPeriod(defaultValue.getEth65TrxAnnouncedBufferingPeriod());
+        .eth65TrxAnnouncedBufferingPeriod(defaultValue.getEth65TrxAnnouncedBufferingPeriod())
+        .txPoolLimitByAccountPercentage(defaultValue.getTxPoolLimitByAccountPercentage());
   }
 
   @Override
@@ -130,7 +167,8 @@ public class TransactionPoolOptionsTest
         .txMessageKeepAliveSeconds(TransactionPoolConfiguration.DEFAULT_TX_MSG_KEEP_ALIVE + 1)
         .eth65TrxAnnouncedBufferingPeriod(
             TransactionPoolConfiguration.ETH65_TRX_ANNOUNCED_BUFFERING_PERIOD.plus(
-                Duration.ofMillis(100)));
+                Duration.ofMillis(100)))
+        .txPoolLimitByAccountPercentage(0.5f);
   }
 
   @Override